lab4/worker/requirements.txt
kafka-python
psycopg2-binary
redis
boto3
